Refactored the endpoint
Improved the implementation structure, separating request structure, response structure, (mime) types and call processing
